+++ drivers/acpi/ospm/busmgr/bm_osl.c
@@ -108,7 +108,7 @@ bm_osl_generate_event (
 	u32			event_data)
 {
 	BM_OSL_EVENT		*event = NULL;
-	u32			flags = 0;
+	unsigned long		flags = 0;
 
 	/* drop event on the floor if no one's listening */
 	if (!event_is_open)
